requirements.txt
Django==4.1.10
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
pymongo==3.12.3
djongo
django-cors-headers